Use beacon person mapper
