+++ src/main/java/uk/gov/mca/beacons/service/dto/AccountHolderDTO.java
@@ -2,7 +2,6 @@ package uk.gov.mca.beacons.service.dto;
 
 import static uk.gov.mca.beacons.service.dto.AccountHolderDTO.Attributes;
 
-import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
