+++ src/main/java/uk/gov/mca/beacons/service/dto/BeaconPersonDTO.java
@@ -1,17 +1,57 @@
 package uk.gov.mca.beacons.service.dto;
 
-import java.util.HashMap;
-import java.util.Map;
+import static uk.gov.mca.beacons.service.dto.BeaconPersonDTO.Attributes;
 
-public class BeaconPersonDTO extends DomainDTO<Map<String, Object>> {
+import com.fasterxml.jackson.annotation.JsonFormat;
+import java.time.LocalDateTime;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
-  public BeaconPersonDTO() {
-    this.attributes = new HashMap<>();
-  }
+public class BeaconPersonDTO extends DomainDTO<Attributes> {
 
   private String type = "beaconPerson";
 
   public String getType() {
     return type;
+  }
+
+  @Getter
+  @Setter
+  @NoArgsConstructor
+  @AllArgsConstructor
+  @Builder
+  public static class Attributes {
+
+    private String fullName;
+    private String email;
+
+    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
+    private LocalDateTime createdDate;
+
+    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
+    private LocalDateTime lastModifiedDate;
+
+    private String telephoneNumber;
+    private String telephoneNumber2;
+    private String alternativeTelephoneNumber;
+    private String alternativeTelephoneNumber2;
+    private String addressLine1;
+    private String addressLine2;
+    private String addressLine3;
+    private String addressLine4;
+    private String townOrCity;
+    private String postcode;
+    private String county;
+    private String country;
+    private String companyName;
+    private String careOf;
+    private String fax;
+    private String isMain;
+    private Integer createUserId;
+    private Integer updateUserId;
+    private Integer versioning;
   }
 }
