+++ src/main/java/uk/gov/mca/beacons/service/mappers/BeaconPersonMapper.java
@@ -1,9 +1,10 @@
 package uk.gov.mca.beacons.service.mappers;
 
-import java.util.HashMap;
-import java.util.Map;
+import static uk.gov.mca.beacons.service.dto.BeaconPersonDTO.Attributes;
+
 import org.springframework.stereotype.Service;
 import uk.gov.mca.beacons.service.dto.BeaconPersonDTO;
+import uk.gov.mca.beacons.service.dto.WrapperDTO;
 import uk.gov.mca.beacons.service.model.BeaconPerson;
 
 @Service
@@ -13,36 +14,79 @@ public class BeaconPersonMapper {
     final var dto = new BeaconPersonDTO();
     dto.setId(domain.getId());
 
-    final Map<String, Object> attributes = new HashMap<>();
-    attributes.put("fullName", domain.getFullName());
-    attributes.put("email", domain.getEmail());
-    attributes.put("telephoneNumber", domain.getTelephoneNumber());
-    attributes.put(
-      "alternativeTelephoneNumber",
-      domain.getAlternativeTelephoneNumber()
-    );
-    attributes.put(
-      "alternativeTelephoneNumber2",
-      domain.getAlternativeTelephoneNumber2()
-    );
-    attributes.put("addressLine1", domain.getAddressLine1());
-    attributes.put("addressLine2", domain.getAddressLine2());
-    attributes.put("addressLine3", domain.getAddressLine3());
-    attributes.put("addressLine4", domain.getAddressLine4());
-    attributes.put("townOrCity", domain.getTownOrCity());
-    attributes.put("county", domain.getCounty());
-    attributes.put("postcode", domain.getPostcode());
-    attributes.put("country", domain.getCountry());
-    attributes.put("companyName", domain.getCompanyName());
-    attributes.put("careOf", domain.getCareOf());
-    attributes.put("fax", domain.getFax());
-    attributes.put("createUserId", domain.getCreateUserId());
-    attributes.put("updateUserId", domain.getUpdateUserId());
-    attributes.put("createdDate", domain.getCreatedDate());
-    attributes.put("lastModifiedDate", domain.getLastModifiedDate());
-    attributes.put("versioning", domain.getVersioning());
+    final Attributes attributes = Attributes
+      .builder()
+      .fullName(domain.getFullName())
+      .email(domain.getEmail())
+      .createdDate(domain.getCreatedDate())
+      .lastModifiedDate(domain.getLastModifiedDate())
+      .telephoneNumber(domain.getTelephoneNumber())
+      .telephoneNumber2(domain.getTelephoneNumber2())
+      .alternativeTelephoneNumber(domain.getAlternativeTelephoneNumber())
+      .alternativeTelephoneNumber2(domain.getAlternativeTelephoneNumber2())
+      .addressLine1(domain.getAddressLine1())
+      .addressLine2(domain.getAddressLine2())
+      .addressLine3(domain.getAddressLine3())
+      .addressLine4(domain.getAddressLine4())
+      .townOrCity(domain.getTownOrCity())
+      .postcode(domain.getPostcode())
+      .county(domain.getCounty())
+      .country(domain.getCountry())
+      .companyName(domain.getCompanyName())
+      .careOf(domain.getCareOf())
+      .fax(domain.getFax())
+      .isMain(domain.getIsMain())
+      .createUserId(domain.getCreateUserId())
+      .updateUserId(domain.getUpdateUserId())
+      .versioning(domain.getVersioning())
+      .build();
     dto.setAttributes(attributes);
 
     return dto;
+  }
+
+  public BeaconPerson fromDTO(BeaconPersonDTO personDTO) {
+    final var beaconPerson = new BeaconPerson();
+    final var attributes = personDTO.getAttributes();
+
+    beaconPerson.setEmail(attributes.getEmail());
+    beaconPerson.setFullName(attributes.getFullName());
+    beaconPerson.setCreatedDate(attributes.getCreatedDate());
+    beaconPerson.setLastModifiedDate(attributes.getLastModifiedDate());
+    beaconPerson.setTelephoneNumber(attributes.getTelephoneNumber());
+    beaconPerson.setTelephoneNumber2(attributes.getTelephoneNumber2());
+    beaconPerson.setAlternativeTelephoneNumber(
+      attributes.getAlternativeTelephoneNumber()
+    );
+    beaconPerson.setAlternativeTelephoneNumber2(
+      attributes.getAlternativeTelephoneNumber2()
+    );
+    beaconPerson.setAddressLine1(attributes.getAddressLine1());
+    beaconPerson.setAddressLine2(attributes.getAddressLine2());
+    beaconPerson.setAddressLine3(attributes.getAddressLine3());
+    beaconPerson.setAddressLine4(attributes.getAddressLine4());
+    beaconPerson.setTownOrCity(attributes.getTownOrCity());
+    beaconPerson.setPostcode(attributes.getPostcode());
+    beaconPerson.setCounty(attributes.getCounty());
+    beaconPerson.setCountry(attributes.getCountry());
+    beaconPerson.setCompanyName(attributes.getCompanyName());
+    beaconPerson.setCareOf(attributes.getCareOf());
+    beaconPerson.setFax(attributes.getFax());
+    beaconPerson.setIsMain(attributes.getIsMain());
+    beaconPerson.setCreateUserId(attributes.getCreateUserId());
+    beaconPerson.setUpdateUserId(attributes.getUpdateUserId());
+    beaconPerson.setVersioning(attributes.getVersioning());
+
+    return beaconPerson;
+  }
+
+  public final WrapperDTO<BeaconPersonDTO> toWrapperDTO(
+    BeaconPerson beaconPerson
+  ) {
+    final WrapperDTO<BeaconPersonDTO> wrapperDTO = new WrapperDTO<>();
+
+    wrapperDTO.setData(toDTO(beaconPerson));
+
+    return wrapperDTO;
   }
 }
