+++ src/main/resources/db/migration/V1.16__Add_Additional_Beacon_Person_Fields_For_Owners.sql
