+++ src/test/resources/fixtures/createOwnerResponseWithDatesSpecified.json
@@ -1,12 +1,12 @@
 {
   "meta": {},
   "data": {
-    "type": "owner",
+    "type": "beaconPerson",
     "attributes": {
       "email": "testy@mctestface.com",
       "fullName": "Tesy McTestface",
-      "createdDate": "1985-10-21T00:00:00",
-      "lastModifiedDate": "1955-10-21T00:00:00",
+      "createdDate": "1985-10-21 00:00:00",
+      "lastModifiedDate": "1955-10-21 00:00:00",
       "telephoneNumber": "01178 657123",
       "telephoneNumber2": "01178 657124",
       "alternativeTelephoneNumber": "01178 657124",
